challenge 34
Display the following message: If the user enters 1 (square), then it should ask them for the length of one of its sides and display the area. If they select 2(triangel), it should ask for the base and height of the triangle and display the area. If they type in anything else, it should give them a suitable error message.