challenge 31
You are going to write a program that tests the compatibility between two people. We're going to use the super scientific method recommended by BuzzFeed.To work out the love score between two people:
Take both people's names and check for the number of times the letters in the word TRUE occurs. Then check for the number of times the letters in the word LOVE occurs. Then combine these numbers to make a 2 digit number.