challenge 128
Create a window that will ask the user to enter a number in a text box. When they click on a button it will use the code variable.isdigit() to check to see if it is a whole number. If it is a whole number, add it to a list box, otherwise clear the entry box. Add another button that will clear the list.